requirements.txt
PyQt5
SpeechRecognition
PyAudio
pyttsx3
pyautogui
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
pytz
python-dateutil
requests
openai
